+++ src/config.ts
@@ -1,0 +1,1 @@
+export const PROJECT_ID = 'e0818fff-92ff-4d0c-bbd9-8df3848883ba'