+++ src/graphql/fragments/service.ts
@@ -1,0 +1,9 @@
+import { gql } from "@apollo/client";
+
+export const SERVICE_FIELDS = gql`
+  fragment ServiceFields on Service {
+    name
+    icon
+    id
+  }
+`;