+++ src/graphql/mutations/restartDeployment.ts
@@ -1,0 +1,6 @@
+import { gql } from "@apollo/client"
+
+export const RESTART_DEPLOYMENT = gql`
+  mutation RestartDeployment($id: String!) {
+    deploymentRestart(id: $id)
+  }`