+++ src/graphql/mutations/stopDeployment.ts
@@ -1,0 +1,7 @@
+import { gql } from "@apollo/client"
+
+export const STOP_DEPLOYMENT = gql`
+  mutation StopDeployment($id: String!) {
+    deploymentStop(id: $id)
+  }
+`