+++ src/graphql/queries/getDeployments.ts
@@ -4,8 +4,8 @@ import { DEPLOYMENT_FIELDS } from '../fragments/deployment';
 export const GET_DEPLOYMENTS = gql`
   ${DEPLOYMENT_FIELDS}
 
-  query GetDeployments($projectId: String) {
-    deployments(input: { projectId: $projectId }) {
+  query GetDeployments($input: DeploymentListInput!) {
+    deployments(input: $input) {
       edges {
         node {
           ...DeploymentFields