+++ src/graphql/queries/getProject.ts
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client"
+
+export const GET_PROJECT = gql`
+  query GetProject($id: String!) {
+    project(id: $id) {
+      id
+      name
+    }
+  }
+`