+++ src/lib/types.ts
@@ -1,0 +1,3 @@
+import { GetServicesQuery } from "@/graphql/generated/graphql";
+
+export type Service = GetServicesQuery['project']['services']['edges'][number]['node']